Torch Callback

This wraps a CallbackSet and annotates it with metadata, most importantly a ParamSet. The callback is created for the given parameter values by calling the $generate() method.

This class is usually used to configure the callback of a torch learner, e.g. when constructing a learner of in a ModelDescriptor.

For a list of available callbacks, see mlr3torch_callbacks. To conveniently retrieve a TorchCallback, use t_clbk().

Parameters

Defined by the constructor argument param_set. If no parameter set is provided during construction, the parameter set is constructed by creating a parameter for each argument of the wrapped loss function, where the parametes are then of type ParamUty.

See also

Super class

mlr3torch::TorchDescriptor -> TorchCallback

Methods

Public methods

TorchCallback$new()

TorchCallback$clone()

Inherited methods

mlr3torch::TorchDescriptor$generate()

mlr3torch::TorchDescriptor$help()

mlr3torch::TorchDescriptor$print()

Method new()

Creates a new instance of this R6 class.

Usage

TorchCallback$new(
  callback_generator,
  param_set = NULL,
  id = NULL,
  label = NULL,
  packages = NULL,
  man = NULL
)

Arguments

callback_generator

(R6ClassGenerator)
The class generator for the callback that is being wrapped.

param_set

(ParamSet or NULL)
The parameter set. If NULL (default) it is inferred from callback_generator.

id

(character(1))
The id for of the new object.

label

(character(1))
Label for the new instance.

packages

(character())
The R packages this object depends on.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Method clone()

The objects of this class are cloneable with this method.

Usage

TorchCallback$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a new torch callback from an existing callback set
torch_callback = TorchCallback$new(CallbackSetCheckpoint)
# The parameters are inferred
torch_callback$param_set
#> <ParamSet>
#>      id    class lower upper nlevels        default value
#> 1: path ParamUty    NA    NA     Inf <NoDefault[3]>      
#> 2: freq ParamUty    NA    NA     Inf <NoDefault[3]>      

# Retrieve a torch callback from the dictionary
torch_callback = t_clbk("checkpoint",
  path = tempfile(), freq = 1
)
torch_callback
#> <TorchCallback:checkpoint> Checkpoint
#> * Generator: CallbackSetCheckpoint
#> * Parameters: path=/tmp/RtmpXrT69R/file19a4ce21dba, freq=1
#> * Packages: mlr3torch,torch
torch_callback$label
#> [1] "Checkpoint"
torch_callback$id
#> [1] "checkpoint"

# open the help page of the wrapped callback set
# torch_callback$help()

# Create the callback set
callback = torch_callback$generate()
callback
#> <CallbackSetCheckpoint>
#>   Inherits from: <CallbackSet>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     ctx: NULL
#>     freq: 1
#>     initialize: function (path, freq) 
#>     on_end: function () 
#>     on_epoch_end: function () 
#>     path: /tmp/RtmpXrT69R/file19a4ce21dba
#>   Private:
#>     deep_clone: function (name, value) 
# is the same as
CallbackSetCheckpoint$new(
  path = tempfile(), freq = 1
)
#> <CallbackSetCheckpoint>
#>   Inherits from: <CallbackSet>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     ctx: NULL
#>     freq: 1
#>     initialize: function (path, freq) 
#>     on_end: function () 
#>     on_epoch_end: function () 
#>     path: /tmp/RtmpXrT69R/file19a443ed4a5e
#>   Private:
#>     deep_clone: function (name, value) 

# Use in a learner
learner = lrn("regr.mlp", callbacks = t_clbk("checkpoint"))
# the parameters of the callback are added to the learner's parameter set
learner$param_set
#> <ParamSetCollection>
#>                     id    class lower upper nlevels        default        value
#>  1:             layers ParamInt 0e+00   Inf     Inf <NoDefault[3]>             
#>  2:           d_hidden ParamInt 1e+00   Inf     Inf <NoDefault[3]>             
#>  3:                  p ParamDbl 0e+00 1e+00     Inf <NoDefault[3]>          0.5
#>  4:         activation ParamUty    NA    NA     Inf <NoDefault[3]> <nn_relu[1]>
#>  5:    activation_args ParamUty    NA    NA     Inf <NoDefault[3]>    <list[0]>
#>  6:         batch_size ParamInt 1e+00   Inf     Inf <NoDefault[3]>             
#>  7:             epochs ParamInt 0e+00   Inf     Inf <NoDefault[3]>             
#>  8:             device ParamFct    NA    NA      12 <NoDefault[3]>         auto
#>  9:     measures_train ParamUty    NA    NA     Inf <NoDefault[3]>    <list[0]>
#> 10:     measures_valid ParamUty    NA    NA     Inf <NoDefault[3]>    <list[0]>
#> 11:          drop_last ParamLgl    NA    NA       2 <NoDefault[3]>        FALSE
#> 12:            shuffle ParamLgl    NA    NA       2 <NoDefault[3]>         TRUE
#> 13:        num_threads ParamInt 1e+00   Inf     Inf <NoDefault[3]>            1
#> 14:               seed ParamInt  -Inf   Inf     Inf <NoDefault[3]>       random
#> 15:             opt.lr ParamDbl 0e+00   Inf     Inf          0.001             
#> 16:          opt.betas ParamUty    NA    NA     Inf    0.900,0.999             
#> 17:            opt.eps ParamDbl 1e-16 1e-04     Inf          1e-08             
#> 18:   opt.weight_decay ParamDbl 0e+00 1e+00     Inf              0             
#> 19:        opt.amsgrad ParamLgl    NA    NA       2          FALSE             
#> 20:     loss.reduction ParamFct    NA    NA       2           mean             
#> 21: cb.checkpoint.path ParamUty    NA    NA     Inf <NoDefault[3]>             
#> 22: cb.checkpoint.freq ParamInt 1e+00   Inf     Inf <NoDefault[3]>             
#>                     id    class lower upper nlevels        default        value