Torch Ingress Token

This function creates an S3 class of class "TorchIngressToken", which is an internal data structure. It contains the (meta-)information of how a batch is generated from a Task and fed into an entry point of the neural network. It is stored as the ingress field in a ModelDescriptor.

TorchIngressToken(features, batchgetter, shape)

Arguments

features

(character)
Features on which the batchgetter will operate.

batchgetter

(function)
Function with two arguments: data and device. This function is given the output of Task$data(rows = batch_indices, cols = features) and it should produce a tensor of shape shape_out.

shape

(integer)
Shape that batchgetter will produce. Batch-dimension should be included as NA.

Value

TorchIngressToken object.