Torch Loss

This wraps a torch::nn_loss and annotates it with metadata, most importantly a ParamSet. The loss function is created for the given parameter values by calling the $generate() method.

This class is usually used to configure the loss function of a torch learner, e.g. when construcing a learner or in a ModelDescriptor.

For a list of available losses, see mlr3torch_losses. Items from this dictionary can be retrieved using t_loss().

Parameters

Defined by the constructor argument param_set. If no parameter set is provided during construction, the parameter set is constructed by creating a parameter for each argument of the wrapped loss function, where the parametes are then of type ParamUty.

See also

Super class

mlr3torch::TorchDescriptor -> TorchLoss

Public fields

task_types

(character())
The task types this loss supports.

Methods

Public methods

TorchLoss$new()

TorchLoss$print()

TorchLoss$clone()

Inherited methods

mlr3torch::TorchDescriptor$generate()

mlr3torch::TorchDescriptor$help()

Method new()

Creates a new instance of this R6 class.

Usage

TorchLoss$new(
  torch_loss,
  task_types,
  param_set = NULL,
  id = NULL,
  label = NULL,
  packages = NULL,
  man = NULL
)

Arguments

torch_loss

(nn_loss)
The loss module.

task_types

(character())
The task types supported by this loss.

param_set

(ParamSet or NULL)
The parameter set. If NULL (default) it is inferred from torch_loss.

id

(character(1))
The id for of the new object.

label

(character(1))
Label for the new instance.

packages

(character())
The R packages this object depends on.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Method print()

Prints the object

Usage

TorchLoss$print(...)

Arguments

...

any

Method clone()

The objects of this class are cloneable with this method.

Usage

TorchLoss$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a new torch loss
torch_loss = TorchLoss$new(torch_loss = nn_mse_loss, task_types = "regr")
torch_loss
#> <TorchLoss:nn_mse_loss> nn_mse_loss
#> * Generator: nn_mse_loss
#> * Parameters: list()
#> * Packages: torch
#> * Task Types: regr
# the parameters are inferred
torch_loss$param_set
#> <ParamSet>
#>           id    class lower upper nlevels
#>       <char>   <char> <num> <num>   <num>
#> 1: reduction ParamUty    NA    NA     Inf
#>                                                                                      default
#>                                                                                       <list>
#> 1: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#>     value
#>    <list>
#> 1:       

# Retrieve a loss from the dictionary:
torch_loss = t_loss("mse", reduction = "mean")
# is the same as
torch_loss
#> <TorchLoss:mse> Mean Squared Error
#> * Generator: nn_mse_loss
#> * Parameters: reduction=mean
#> * Packages: torch
#> * Task Types: regr
torch_loss$param_set
#> <ParamSet>
#>           id    class lower upper nlevels default  value
#>       <char>   <char> <num> <num>   <int>  <list> <list>
#> 1: reduction ParamFct    NA    NA       2    mean   mean
torch_loss$label
#> [1] "Mean Squared Error"
torch_loss$task_types
#> [1] "regr"
torch_loss$id
#> [1] "mse"

# Create the loss function
loss_fn = torch_loss$generate()
loss_fn
#> An `nn_module` containing 0 parameters.
# Is the same as
nn_mse_loss(reduction = "mean")
#> An `nn_module` containing 0 parameters.

# open the help page of the wrapped loss function
# torch_loss$help()

# Use in a learner
learner = lrn("regr.mlp", loss = t_loss("mse"))
# The parameters of the loss are added to the learner's parameter set
learner$param_set
#> <ParamSetCollection>
#>                   id    class lower upper nlevels
#>               <char>   <char> <num> <num>   <num>
#>  1:          neurons ParamUty    NA    NA     Inf
#>  2:                p ParamDbl 0e+00 1e+00     Inf
#>  3:       activation ParamUty    NA    NA     Inf
#>  4:  activation_args ParamUty    NA    NA     Inf
#>  5:            shape ParamUty    NA    NA     Inf
#>  6:       batch_size ParamInt 1e+00   Inf     Inf
#>  7:           epochs ParamInt 0e+00   Inf     Inf
#>  8:           device ParamFct    NA    NA      12
#>  9:   measures_train ParamUty    NA    NA     Inf
#> 10:   measures_valid ParamUty    NA    NA     Inf
#> 11:        drop_last ParamLgl    NA    NA       2
#> 12:          shuffle ParamLgl    NA    NA       2
#> 13:      num_threads ParamInt 1e+00   Inf     Inf
#> 14:             seed ParamInt  -Inf   Inf     Inf
#> 15:           opt.lr ParamDbl 0e+00   Inf     Inf
#> 16:        opt.betas ParamUty    NA    NA     Inf
#> 17:          opt.eps ParamDbl 1e-16 1e-04     Inf
#> 18: opt.weight_decay ParamDbl 0e+00 1e+00     Inf
#> 19:      opt.amsgrad ParamLgl    NA    NA       2
#> 20:   loss.reduction ParamFct    NA    NA       2
#>                   id    class lower upper nlevels
#>                                                                                       default
#>                                                                                        <list>
#>  1: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#>  2: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#>  3: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#>  4: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#>  5: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#>  6: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#>  7: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#>  8: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#>  9: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#> 10: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#> 11: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#> 12: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#> 13: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#> 14: <NoDefault>\n  Public:\n    clone: function (deep = FALSE) \n    initialize: function () 
#> 15:                                                                                     0.001
#> 16:                                                                               0.900,0.999
#> 17:                                                                                     1e-08
#> 18:                                                                                         0
#> 19:                                                                                     FALSE
#> 20:                                                                                      mean
#>                                                                                       default
#>            value
#>           <list>
#>  1:             
#>  2:          0.5
#>  3: <nn_relu[1]>
#>  4:    <list[0]>
#>  5:             
#>  6:             
#>  7:             
#>  8:         auto
#>  9:    <list[0]>
#> 10:    <list[0]>
#> 11:        FALSE
#> 12:         TRUE
#> 13:            1
#> 14:       random
#> 15:             
#> 16:             
#> 17:             
#> 18:             
#> 19:             
#> 20:             
#>            value