Torch Optimizer

This wraps a torch::torch_optimizer_generatora and annotates it with metadata, most importantly a ParamSet. The optimizer is created for the given parameter values by calling the $generate() method.

This class is usually used to configure the optimizer of a torch learner, e.g. when construcing a learner or in a ModelDescriptor.

For a list of available optimizers, see mlr3torch_optimizers. Items from this dictionary can be retrieved using t_opt().

Parameters

Defined by the constructor argument param_set. If no parameter set is provided during construction, the parameter set is constructed by creating a parameter for each argument of the wrapped loss function, where the parametes are then of type ParamUty.

See also

Super class

mlr3torch::TorchDescriptor -> TorchOptimizer

Methods

Public methods

TorchOptimizer$new()

TorchOptimizer$generate()

TorchOptimizer$clone()

Inherited methods

mlr3torch::TorchDescriptor$help()

mlr3torch::TorchDescriptor$print()

Method new()

Creates a new instance of this R6 class.

Usage

TorchOptimizer$new(
  torch_optimizer,
  param_set = NULL,
  id = NULL,
  label = NULL,
  packages = NULL,
  man = NULL
)

Arguments

torch_optimizer

(torch_optimizer_generator)
The torch optimizer.

param_set

(ParamSet or NULL)
The parameter set. If NULL (default) it is inferred from torch_optimizer.

id

(character(1))
The id for of the new object.

label

(character(1))
Label for the new instance.

packages

(character())
The R packages this object depends on.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Method generate()

Instantiates the optimizer.

Usage

TorchOptimizer$generate(params)

Arguments

params

The parameters of the network.

Method clone()

The objects of this class are cloneable with this method.

Usage

TorchOptimizer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a new torch loss
torch_opt = TorchOptimizer$new(optim_adam, label = "adam")
torch_opt
#> <TorchOptimizer:optim_adam> adam
#> * Generator: optim_adam
#> * Parameters: list()
#> * Packages: torch,mlr3torch
# If the param set is not specified, parameters are inferred but are of class ParamUty
torch_opt$param_set
#> <ParamSet>
#>              id    class lower upper nlevels        default value
#> 1:           lr ParamUty    NA    NA     Inf <NoDefault[3]>      
#> 2:        betas ParamUty    NA    NA     Inf <NoDefault[3]>      
#> 3:          eps ParamUty    NA    NA     Inf <NoDefault[3]>      
#> 4: weight_decay ParamUty    NA    NA     Inf <NoDefault[3]>      
#> 5:      amsgrad ParamUty    NA    NA     Inf <NoDefault[3]>      

# open the help page of the wrapped optimizer
# torch_opt$help()

# Retrieve an optimizer from the dictionary
torch_opt = t_opt("sgd", lr = 0.1)
torch_opt
#> <TorchOptimizer:sgd> Stochastic Gradient Descent
#> * Generator: optim_sgd
#> * Parameters: lr=0.1
#> * Packages: torch,mlr3torch
torch_opt$param_set
#> <ParamSet>
#>              id    class lower upper nlevels        default value
#> 1:           lr ParamDbl     0   Inf     Inf <NoDefault[3]>   0.1
#> 2:     momentum ParamDbl     0     1     Inf              0      
#> 3:    dampening ParamDbl     0     1     Inf              0      
#> 4: weight_decay ParamDbl     0     1     Inf              0      
#> 5:     nesterov ParamLgl    NA    NA       2          FALSE      
torch_opt$label
#> [1] "Stochastic Gradient Descent"
torch_opt$id
#> [1] "sgd"

# Create the optimizer for a network
net = nn_linear(10, 1)
opt = torch_opt$generate(net$parameters)

# is the same as
optim_sgd(net$parameters, lr = 0.1)
#> <optim_sgd>
#>   Inherits from: <torch_optimizer>
#>   Public:
#>     add_param_group: function (param_group) 
#>     clone: function (deep = FALSE) 
#>     defaults: list
#>     initialize: function (params, lr = optim_required(), momentum = 0, dampening = 0, 
#>     load_state_dict: function (state_dict, ..., .refer_to_state_dict = FALSE) 
#>     param_groups: list
#>     state: State, R6
#>     state_dict: function () 
#>     step: function (closure = NULL) 
#>     zero_grad: function () 
#>   Private:
#>     step_helper: function (closure, loop_fun) 

# Use in a learner
learner = lrn("regr.mlp", optimizer = t_opt("sgd"))
# The parameters of the optimizer are added to the learner's parameter set
learner$param_set
#> <ParamSetCollection>
#>                   id    class lower upper nlevels        default        value
#>  1:           layers ParamInt     0   Inf     Inf <NoDefault[3]>             
#>  2:         d_hidden ParamInt     1   Inf     Inf <NoDefault[3]>             
#>  3:                p ParamDbl     0     1     Inf <NoDefault[3]>          0.5
#>  4:       activation ParamUty    NA    NA     Inf <NoDefault[3]> <nn_relu[1]>
#>  5:  activation_args ParamUty    NA    NA     Inf <NoDefault[3]>    <list[0]>
#>  6:       batch_size ParamInt     1   Inf     Inf <NoDefault[3]>             
#>  7:           epochs ParamInt     0   Inf     Inf <NoDefault[3]>             
#>  8:           device ParamFct    NA    NA      11 <NoDefault[3]>         auto
#>  9:   measures_train ParamUty    NA    NA     Inf <NoDefault[3]>    <list[0]>
#> 10:   measures_valid ParamUty    NA    NA     Inf <NoDefault[3]>    <list[0]>
#> 11:        drop_last ParamLgl    NA    NA       2 <NoDefault[3]>        FALSE
#> 12:          shuffle ParamLgl    NA    NA       2 <NoDefault[3]>         TRUE
#> 13:      num_threads ParamInt     1   Inf     Inf <NoDefault[3]>            1
#> 14:             seed ParamInt  -Inf   Inf     Inf <NoDefault[3]>       random
#> 15:           opt.lr ParamDbl     0   Inf     Inf <NoDefault[3]>             
#> 16:     opt.momentum ParamDbl     0     1     Inf              0             
#> 17:    opt.dampening ParamDbl     0     1     Inf              0             
#> 18: opt.weight_decay ParamDbl     0     1     Inf              0             
#> 19:     opt.nesterov ParamLgl    NA    NA       2          FALSE             
#> 20:   loss.reduction ParamFct    NA    NA       2           mean