Autotest for PipeOpTorch

This tests a PipeOpTorch that is embedded in a Graph. Note that the id of the PipeOp to test should be the default id. It tests whether:

the output tensor(s) have the correct shape(s)

the generated module has the intended class

the parameters are correctly implemented

Some other basic checks

autotest_pipeop_torch(
  graph,
  id,
  task,
  module_class = id,
  exclude_args = character(0)
)

Arguments

graph

(Graph)
The graph that contains the PipeOpTorch to be tested.

id

(character(1))
The id of the PipeOpTorch to be tested.

task

(Task)
The task with which the PipeOpTorch will be tested.

module_class

(character(1))
The module class that is expected from the generated module of the PipeOpTorch. Default is the PipeOps id.

exclude_args

(character())
We check that the arguments of the module's forward function match the input channels except for those values specified as exclude_args.

Value

TRUE if the autotest passes, errs otherwise.

Details

TODO