Batchgetter for Numeric Data

Converts a data frame of numeric data into a float tensor by calling as.matrix(). No input checks are performed

Usage

batchgetter_num(data, device, ...)

Arguments

data

(data.table())
data.table to be converted to a tensor.

device

(character(1))
The device on which the tensor should be created.

...

(any)
Unused.