Base Class for Callbacks

Base class from which callbacks should inherit (see section Inheriting). A callback set is a collection of functions that are executed at different stages of the training loop. They can be used to gain more control over the training process of a neural network without having to write everything from scratch.

When used a in torch learner, the CallbackSet is wrapped in a TorchCallback. The latters parameter set represents the arguments of the CallbackSet's $initialize() method.

Inheriting

For each available stage (see section Stages) a public method $on_<stage>() can be defined. The evaluation context (a ContextTorch) can be accessed via self$ctx, which contains the current state of the training loop. This context is assigned at the beginning of the training loop and removed afterwards. Different stages of a callback can communicate with each other by assigning values to $self.

For creating custom callbacks, the function torch_callback() is recommended, which creates a CallbackSet and then wraps it in a TorchCallback. To create a CallbackSet the convenience function callback_set() can be used. These functions perform checks such as that the stages are not accidentally misspelled.

Stages

begin :: Run before the training loop begins.

epoch_begin :: Run he beginning of each epoch.

before_validation :: Run before each validation loop.

batch_begin :: Run before the forward call.

after_backward :: Run after the backward call.

batch_end :: Run after the optimizer step.

batch_valid_begin :: Run before the forward call in the validation loop.

batch_valid_end :: Run after the forward call in the validation loop.

epoch_end :: Run at the end of each epoch.

end :: Run at last, using on.exit().

See also

Public fields

ctx

(ContextTorch or NULL)
The evaluation context for the callback. This field should always be NULL except during the $train() call of the torch learner.

Active bindings

stages

(character())
The active stages of this callback set.

Methods

Public methods

CallbackSet$print()

CallbackSet$clone()

Method print()

Prints the object.

Usage

CallbackSet$print(...)

Arguments

...

(any)
Currently unused.

Method clone()

The objects of this class are cloneable with this method.

Usage

CallbackSet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.