Checkpoint Callback

Saves the model during training.

See also

Super class

mlr3torch::CallbackSet -> CallbackSetCheckpoint

Methods

Public methods

CallbackSetCheckpoint$new()

CallbackSetCheckpoint$on_epoch_end()

CallbackSetCheckpoint$on_end()

CallbackSetCheckpoint$clone()

Method new()

Creates a new instance of this R6 class.

Usage

CallbackSetCheckpoint$new(path, freq)

Arguments

path

(character(1))
The path to a folder where the models are saved. This path must not exist before.

freq

(integer(1))
The frequency how often the model is saved (epoch frequency).

Method on_epoch_end()

Saves the network state dict.

Usage

CallbackSetCheckpoint$on_epoch_end()

Method on_end()

Saves the final network.

Usage

CallbackSetCheckpoint$on_end()

Method clone()

The objects of this class are cloneable with this method.

Usage

CallbackSetCheckpoint$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.