Progress Callback

Prints a progress bar and the metrics for training and validation.

See also

Super class

mlr3torch::CallbackSet -> CallbackSetProgress

Methods

Public methods

CallbackSetProgress$on_epoch_begin()

CallbackSetProgress$on_batch_end()

CallbackSetProgress$on_before_valid()

CallbackSetProgress$on_batch_valid_end()

CallbackSetProgress$on_epoch_end()

CallbackSetProgress$clone()

Inherited methods

mlr3torch::CallbackSet$load_state_dict()

mlr3torch::CallbackSet$print()

mlr3torch::CallbackSet$state_dict()

Method on_epoch_begin()

Initializes the progress bar for training.

Usage

CallbackSetProgress$on_epoch_begin()

Method on_batch_end()

Increments the training progress bar.

Usage

CallbackSetProgress$on_batch_end()

Method on_before_valid()

Creates the progress bar for validation.

Usage

CallbackSetProgress$on_before_valid()

Method on_batch_valid_end()

Increments the validation progress bar.

Usage

CallbackSetProgress$on_batch_valid_end()

Method on_epoch_end()

Prints a summary of the training and validation process.

Usage

CallbackSetProgress$on_epoch_end()

Method clone()

The objects of this class are cloneable with this method.

Usage

CallbackSetProgress$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

task = tsk("iris")

learner = lrn("classif.mlp", epochs = 10, batch_size = 1, 
  callbacks = t_clbk("progress"), validate = 0.1)
learner$param_set$set_values(
measures_train = msrs(c("classif.acc", "classif.ce")),
  measures_valid = msr("classif.ce")
)

learner$train(task)
#> Epoch 1
#> 
#> [Summary epoch 1]
#> ------------------
#> Measures (Train):
#>  * classif.acc = 0.33
#>  * classif.ce = 0.67
#> Measures (Valid):
#>  * classif.ce = 0.60
#> Epoch 2
#> 
#> [Summary epoch 2]
#> ------------------
#> Measures (Train):
#>  * classif.acc = 0.42
#>  * classif.ce = 0.58
#> Measures (Valid):
#>  * classif.ce = 0.20
#> Epoch 3
#> 
#> [Summary epoch 3]
#> ------------------
#> Measures (Train):
#>  * classif.acc = 0.50
#>  * classif.ce = 0.50
#> Measures (Valid):
#>  * classif.ce = 0.20
#> Epoch 4
#> 
#> [Summary epoch 4]
#> ------------------
#> Measures (Train):
#>  * classif.acc = 0.54
#>  * classif.ce = 0.46
#> Measures (Valid):
#>  * classif.ce = 0.33
#> Epoch 5
#> 
#> [Summary epoch 5]
#> ------------------
#> Measures (Train):
#>  * classif.acc = 0.64
#>  * classif.ce = 0.36
#> Measures (Valid):
#>  * classif.ce = 0.40
#> Epoch 6
#> 
#> [Summary epoch 6]
#> ------------------
#> Measures (Train):
#>  * classif.acc = 0.61
#>  * classif.ce = 0.39
#> Measures (Valid):
#>  * classif.ce = 0.40
#> Epoch 7
#> 
#> [Summary epoch 7]
#> ------------------
#> Measures (Train):
#>  * classif.acc = 0.61
#>  * classif.ce = 0.39
#> Measures (Valid):
#>  * classif.ce = 0.40
#> Epoch 8
#> 
#> [Summary epoch 8]
#> ------------------
#> Measures (Train):
#>  * classif.acc = 0.61
#>  * classif.ce = 0.39
#> Measures (Valid):
#>  * classif.ce = 0.40
#> Epoch 9
#> 
#> [Summary epoch 9]
#> ------------------
#> Measures (Train):
#>  * classif.acc = 0.62
#>  * classif.ce = 0.38
#> Measures (Valid):
#>  * classif.ce = 0.40
#> Epoch 10
#> 
#> [Summary epoch 10]
#> ------------------
#> Measures (Train):
#>  * classif.acc = 0.64
#>  * classif.ce = 0.36
#> Measures (Valid):
#>  * classif.ce = 0.40