Base Class for Torch Callbacks

Base class from which Torch Callbacks should inherit. To create custom callbacks to use in a torch learner use the convenience function torch_callback.

Torch Callbacks can be used to gain more control over the training process of a neural network without having to write everything from scratch. At each stage (see section "Stages") of the training loop, the corresponding on_<stage>(ctx) method is run that takes as argument a ContextTorch which gives access to the relevant objects.

Format

R6Class

Construction

CallbackTorch$new()

Methods

See section Stages. Other methods can be added freely as well.

Inheriting

It is recommended to use the sugar function callback_torch() to create custom callbacks.

Stages

begin :: Run before the training loop begins.

epoch_begin :: Run he beginning of each epoch.

before_validation :: Run before each validation loop.

batch_begin :: Run before the forward call.

after_backward :: Run after the backward call.

batch_end :: Run after the optimizer step.

batch_valid_begin :: Run before the forward call in the validation loop.

batch_valid_end :: Run after the forward call in the validation loop.

epoch_end :: Run at the end of each epoch.

end :: Run at last, using on.exit().

See also