Context for Torch Learner

Context for training a torch learner. This is the - mostly read-only - information callbacks have access to through the argument ctx. For more information on callbacks, see CallbackTorch.

Format

R6Class inheriting from Context.

Construction

ContextTorch$new(learner, task_train, task_valid = NULL, loader_train, loader_valid = NULL, measures_train = NULL, measures_valid = NULL, network, optimizer, loss_fn, total_epochs)

learner :: Learner
The torch learner.

task_train :: Task
The training task.

task_valid :: Task
The validation task. Default is NULL.

loader_train :: torch::dataloader The data loader for training.

loader_valid :: torch::dataloader The data loader for validation.

measures_train :: list() of Measures
Measures used for training. Default is NULL.

measures_valid :: list() of Measures
Measures used for validation. Default is NULL.

network :: torch::nn_module
The torch network.

optimizer :: torch::optimizer
The optimizer.

loss_fn :: torch::nn_module
The loss function.

total_epochs :: integer(1)
The total number of epochs the learner is trained for.

Fields

All objects configured in the initialization, as well as:

last_scores_train :: named list()
The scores from the last training batch. Names are the ids of the training measures.

last_scores_valid :: named list()
The scores from the last validation batch. Names are the ids of the validation measures.

epoch :: integer(1)
The current epoch. The current epoch.

batch :: list() of torch::torch_tensor
The current iteration of the batch.

Methods

There are no methods.

See also