AlexNet Image Classifier

Historic convolutional neural network for image classification.

Dictionary

This Learner can be instantiated using the sugar function lrn():

lrn("classif.alexnet", ...)

Meta Information

Supported task types: 'classif'

Predict Types:

classif: 'response', 'prob'

Feature Types: “lazy_tensor”

Required Packages: mlr3, mlr3torch, torchvision, magick, torch

State

The state is a list with elements:

network :: The trained network.

optimizer :: The optimizer used to train the network.

loss_fn :: The loss used to train the network.

callbacks :: The callbacks used to train the network.

seed :: The actual seed that was / is used for training and prediction.

Parameters

Parameters from LearnerTorchImage and

pretrained :: logical(1)
Whether to use the pretrained model.

References

Krizhevsky, Alex, Sutskever, Ilya, Hinton, E. G (2017). “Imagenet classification with deep convolutional neural networks.” Communications of the ACM, 60(6), 84--90.

See also

Super classes

mlr3::Learner -> mlr3torch::LearnerTorch -> mlr3torch::LearnerTorchImage -> LearnerTorchAlexNet

Methods

Public methods

LearnerTorchAlexNet$new()

LearnerTorchAlexNet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3torch::LearnerTorch$format()

mlr3torch::LearnerTorch$marshal()

mlr3torch::LearnerTorch$print()

mlr3torch::LearnerTorch$unmarshal()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerTorchAlexNet$new(
  task_type,
  optimizer = NULL,
  loss = NULL,
  callbacks = list()
)

Arguments

task_type

(character(1))
The task type, either "classif" or "regr".

optimizer

(TorchOptimizer)
The optimizer to use for training. Per default, adam is used.

loss

(TorchLoss)
The loss used to train the network. Per default, mse is used for regression and cross_entropy for classification.

callbacks

(list() of TorchCallbacks)
The callbacks. Must have unique ids.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerTorchAlexNet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.