Abstract Base Class for a Torch Classification Learner

This base class provides the basic functionality for training and prediction of a neural network. All torch classifiction learners should inherit from the respective subclass.

Format

R6Class inheriting from LearnerClassif, Learner.

Construction

Classification:

LearnerClassifTorch$new(id, optimizer, loss, param_set, properties = c("twoclass", "multiclass"), packages = character(0), predict_types = c("response", "prob"), feature_types, man, label, callbacks = list())

Regression:

LearnerRegrTorch$new(id, optimizer, loss, param_set, properties = character(0), packages = character(0), predict_types = "response", feature_types, man, label, callbacks = list())

id :: character(1)
The id for the object.

optimizer :: TorchOptimizer
The optimizer for the model.

loss :: TorchLoss
The loss for the model.

callbacks:: list() of TorchCallback objects
The callbacks used for training. Must have unique IDs.

param_set :: link{ParamSet}{paradox}
The parameter set.

properties :: character()
The properties for the learner, see mlr_reflections$learner_properties.

packages :: character()
The additional packages on which the learner depends. The packages "torch" and "mlr3torch" are automatically added so do not have to be passed explicitly.

predict_types :: character()
The learner's predict types, see mlr_reflections$learner_predict_types. The default is "response" and "prob".

feature_types :: character()
The feature types the learner can deal with, see mlr_reflections$task_feature_types.

man :: character(1)
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

label :: character(1)
The label for the learner.

State

The state is a list with elements network, optimizer, loss_fn and callbacks.

Parameters

batch_size :: (integer(1))
The batch size.

epochs :: integer(1)
The number of epochs.

device :: character(1)
The device. One of "auto", "cpu", or "cuda".

measures_train :: Measure or list() of Measures. Measures to be evaluated during training.

measures_valid :: Measure or list() of Measures. Measures to be evaluated during validation.

drop_last :: logical(1)
Whether to drop the last batch in each epoch during training. Default is FALSE.

num_threads :: integer(1)
The number of threads (if device is "cpu"). Default is 1.

shuffle :: logical(1)
Whether to shuffle the instances in the dataset. Default is TRUE.

early_stopping_rounds :: integer(1)
How many rounds to wait for early stopping. The default is 0.

seed :: integer(1)
The seed that is used during training. The value seed + 1 is used during prediction. If this is missing (default), a random seed is generated.

Additionally there are the parameters for the optimizer, the loss function and the callbacks. "opt.", "loss." and "cb.<callback id>" respectively.

Inheriting

When inheriting from this class, one should overload two, methods, namely the private$.network(task, param_vals) and the private$.dataset(task, param_vals). The former should construct torch::nn_module object for the given task and parameter values, while the latter is responsible for creating a torch::dataset. Note that the output of this network are expected to be the scores before the application of the final softmax layer.

It is also possible to overwrite the private $.dataloader() method, which otherwise calls $.dataset() and creates a dataloader from that dataset. When doing so, it is important to respect the parameter shuffle, because this method is used to ceate the dataloader for prediction as well.

While it is possible to add parameters by specifying the param_set construction argument, it is currently not possible to change these parameters. Note that none of the parameters provided in param_set can have an id that starts with "loss.", "opt.", or "cb."`, as these are preserved for the dynamically constructed parameters of the optimizer and the loss function.

Fields

Fields inherited from LearnerClassif or LearnerRegr and

network :: (nn_module())
The network (only available after training).

history :: CallbackTorchHistory
The training history.

Methods

Methods inherited from LearnerClassif, Learner.

Internals

A ParamSetCollection is created that combines the param_set from the construction with the default torch parameters, as well as the loss, optimizer and callback parameters (prefixed with "loss.", "opt.", and "cb." respectively.

See also