Abstract Base Class for a Torch Regression Learner

This base class provides the basic functionality for training and prediction of a neural network. All torch regression learners should inherit from the respective subclass.

State

The state is a list with elements network, optimizer, loss_fn and callbacks.

Parameters

batch_size :: (integer(1))
The batch size.

epochs :: integer(1)
The number of epochs.

device :: character(1)
The device. One of "auto", "cpu", or "cuda".

measures_train :: Measure or list() of Measures. Measures to be evaluated during training.

measures_valid :: Measure or list() of Measures. Measures to be evaluated during validation.

drop_last :: logical(1)
Whether to drop the last batch in each epoch during training. Default is FALSE.

num_threads :: integer(1)
The number of threads (if device is "cpu"). Default is 1.

shuffle :: logical(1)
Whether to shuffle the instances in the dataset. Default is TRUE.

early_stopping_rounds :: integer(1)
How many rounds to wait for early stopping. The default is 0.

seed :: integer(1)
The seed that is used during training. The value seed + 1 is used during prediction. If this is missing (default), a random seed is generated.

Additionally there are the parameters for the optimizer, the loss function and the callbacks. They are prefixed with "opt.", "loss." and "cb.<callback id>." respectively.

Inheriting

When inheriting from this class, one should overload two private methods:

.network(task, param_vals)
(Task, list()) -> nn_module
Construct a torch::nn_module object for the given task and parameter values, i.e. the neural network that is trained by the learner. For classification, the output of this network are expected to be the scores before the application of the final softmax layer.

.dataset(task, param_vals)
(Task, list()) -> torch::dataset
Create the dataset for the task. Must at least respect parameters batch_size and shuffle.

It is also possible to overwrite the private .dataloader() method instead of the .dataset() method. Per default, a dataloader is constructed using the output from the .dataset() method.

.dataloader(task, param_vals)
(Task, list()) -> torch::dataloader
Create a dataloader from the task. Needs to respect at least batch_size and shuffle (otherwise predictions are permuted).

While it is possible to add parameters by specifying the param_set construction argument, it is currently not possible to remove existing parameters, i.e. those listed in section Parameters. None of the parameters provided in param_set can have an id that starts with "loss.", "opt.", or "cb."`, as these are preserved for the dynamically constructed parameters of the optimizer, the loss function, and the callbacks.

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrTorch

Active bindings

network

(nn_module())
The network (only available after training).

param_set

(ParamSet)
The parameter set

history

(CallbackTorchHistory)
Shortcut for learner$model$callbacks$history.

Methods

Public methods

LearnerRegrTorch$new()

LearnerRegrTorch$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrTorch$new(
  id,
  optimizer,
  loss,
  param_set,
  properties = character(0),
  packages = character(0),
  predict_types = "response",
  feature_types,
  man,
  label,
  callbacks = list()
)

Arguments

id

(character(1))
The id for of the new object.

optimizer

(TorchOptimizer)
The torch optimizer.

loss

(TorchLoss)
The loss used to train the network.

param_set

(paradox::ParamSet)
The parameter set.

properties

(character())
The properties of the object. See mlr_reflections$learner_properties for available values.

packages

(character())
The R packages this object depends on.

predict_types

(character())
The predict types. See mlr_reflections$learner_predict_types for available values.

feature_types

(character())
The feature types. See mlr_reflections$task_feature_types for available values.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

label

(character(1))
Label for the new instance.

callbacks

(list() of TorchCallbacks)
The callbacks used during training. Must have unique ids. They are executed in the order in which they are provided

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrTorch$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.