Class for Torch Module Wrappers

PipeOpModule wraps an nn_module that is being called during the train phase of this mlr3pipelines::PipeOp. By doing so, this allows to assemble PipeOpModules in a computational mlr3pipelines::Graph that represents a neural network architecture. Such a graph can also be used to create a nn_graph which inherits from nn_module.

In most cases it is easier to create such a network by creating a isomorphic graph consisting of nodes of class PipeOpTorchIngress and PipeOpTorch. This graph will then generate the graph consisting of PipeOpModules during its training phase.

The predict method does currently not serve a meaningful purpose.

Format

R6Class object inheriting from mlr3pipelines::PipeOp.

Construction

PipeOpModule$new(id = "module", module = nn_identity(), inname = "input", outname = "output", param_vals = list(), packages = character(0))

id :: character(1)
The id for the object. The default is "module".

module :: nn_module
The torch module that is being wrapped.

inname :: character()

For a more general introduction on how to create a new learner, see the respective section in the mlr3 book. The names of the input channels.

outname :: character()
The names of the output channels. If this parameter has length 1, the parameter module must return a tensor. Otherwise it must return a list() of tensors of corresponding length.

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

packages` :: named list()
List of packages settings. Default is list().

Input and Output Channels

The number and names of the input and output channels can be set during construction. They input and output "torch_tensor" during training, and NULL during prediction as the prediction phase currently serves no meaningful purpose.

State

The $state is an empty list().

Parameters

No parameters.

Fields

module :: nn_module
The torch module that is called during the training phase of the PipeOpModule.

Methods

Only methods inherited from mlr3pipelines::PipeOp.

Internals

During training, the wrapped nn_module is called with the provided inputs in the order in which the channels are defined. Arguments are not matched by name.

See also

Examples

## creating an PipeOpModule manually

# one input and output channel
po_module = PipeOpModule$new("linear",
  torch::nn_linear(10, 20),
  inname = "input",
  outname = "output"
)
#> Error in cpp_torch_strided(): Lantern is not loaded. Please use `install_torch()` to install additional dependencies.
x = torch::torch_randn(16, 10)
#> Error in cpp_torch_strided(): Lantern is not loaded. Please use `install_torch()` to install additional dependencies.
# This calls the forward function of the wrapped module.
y = po_module$train(list(input = x))
#> Error in eval(expr, envir, enclos): object 'po_module' not found
str(y)
#> Error in eval(expr, envir, enclos): object 'y' not found

# multiple input and output channels
nn_custom = torch::nn_module("nn_custom",
  initialize = function(in_features, out_features) {
    self$lin1 = torch::nn_linear(in_features, out_features)
    self$lin2 = torch::nn_linear(in_features, out_features)
  },
  forward = function(x, z) {
    list(out1 = self$lin1(x), out2 = torch::nnf_relu(self$lin2(z)))
  }
)

module = nn_custom(3, 2)
#> Error in cpp_torch_strided(): Lantern is not loaded. Please use `install_torch()` to install additional dependencies.
po_module = PipeOpModule$new(
  "custom",
  module,
  inname = c("x", "z"),
  outname = c("out1", "out2")
)
#> Error in eval(expr, envir, enclos): object 'module' not found
x = torch::torch_randn(1, 3)
#> Error in cpp_torch_strided(): Lantern is not loaded. Please use `install_torch()` to install additional dependencies.
z = torch::torch_randn(1, 3)
#> Error in cpp_torch_strided(): Lantern is not loaded. Please use `install_torch()` to install additional dependencies.
out = po_module$train(list(x = x, z = z))
#> Error in eval(expr, envir, enclos): object 'po_module' not found
str(out)
#> Error in eval(expr, envir, enclos): object 'out' not found

# How a PipeOpModule is usually generated
graph = po("torch_ingress_num") %>>% po("nn_linear", out_features = 10L)
result = graph$train(tsk("iris"))
#> Error in cpp_torch_strided(): Lantern is not loaded. Please use `install_torch()` to install additional dependencies.
#> This happened PipeOp nn_linear's $train()
# The PipeOpTorchLinear generates a PipeOpModule and adds it to a new (module) graph
result[[1]]$graph
#> Error in eval(expr, envir, enclos): object 'result' not found
linear_module = result[[1L]]$graph$pipeops$nn_linear
#> Error in eval(expr, envir, enclos): object 'result' not found
linear_module
#> Error in eval(expr, envir, enclos): object 'linear_module' not found
formalArgs(linear_module$module)
#> Error in eval(expr, envir, enclos): object 'linear_module' not found
linear_module$input$name
#> Error in eval(expr, envir, enclos): object 'linear_module' not found