Layer Normalization

Applies Layer Normalization for last certain number of dimensions.

State

The state is the value calculated by the public method $shapes_out().

Credit

Part of this documentation have been copied or adapted from the documentation of torch.

Input and Output Channels

One input channel called "input" and one output channel called "output". For an explanation see PipeOpTorch.

Parameters

dims :: integer(1)
The number of dimensions over which will be normalized (starting from the last dimension).

elementwise_affine :: logical(1)
Whether to learn affine-linear parameters initialized to 1 for weights and to 0 for biases. The default is TRUE.

eps :: numeric(1)
A value added to the denominator for numerical stability.

Internals

Calls torch::nn_layer_norm() when trained. The parameter normalized_shape is inferred as the dimensions of the last dims dimensions of the input shape.

See also

Super classes

mlr3pipelines::PipeOp -> mlr3torch::PipeOpTorch -> PipeOpTorchLayerNorm

Methods

Public methods

PipeOpTorchLayerNorm$new()

PipeOpTorchLayerNorm$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

mlr3torch::PipeOpTorch$shapes_out()

Method new()

Creates a new instance of this R6 class.

Usage

PipeOpTorchLayerNorm$new(id = "nn_layer_norm", param_vals = list())

Arguments

id

(character(1))
Identifier of the resulting object.

param_vals

(list())
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction.

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpTorchLayerNorm$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Construct the PipeOp
pipeop = po("nn_layer_norm", dims = 1)
pipeop
#> PipeOp: <nn_layer_norm> (not trained)
#> values: <dims=1>
#> Input channels <name [train type, predict type]>:
#>   input [ModelDescriptor,Task]
#> Output channels <name [train type, predict type]>:
#>   output [ModelDescriptor,Task]
# The available parameters
pipeop$param_set
#> <ParamSet:nn_layer_norm>
#>                    id    class lower upper nlevels        default value
#> 1:               dims ParamInt     1   Inf     Inf <NoDefault[3]>     1
#> 2: elementwise_affine ParamLgl    NA    NA       2           TRUE      
#> 3:                eps ParamDbl     0   Inf     Inf          1e-05