Merge by Product

Calculates the product of all input tensors.

Input and Output Channels

One input channel called "input" and one output channel called "output". For an explanation see PipeOpTorch.

PipeOpTorchMerges has either a vararg input channel if the constructor argument innum is not set, or input channels "input1", ..., "input<innum>". There is one output channel "output". For an explanation see PipeOpTorch.

State

The state is the value calculated by the public method $shapes_out().

Credit

Part of this documentation have been copied or adapted from the documentation of torch.

Parameters

No parameters.

Internals

Calls nn_merge_prod() when trained.

See also

Super classes

mlr3pipelines::PipeOp -> mlr3torch::PipeOpTorch -> mlr3torch::PipeOpTorchMerge -> PipeOpTorchMergeProd

Methods

Public methods

PipeOpTorchMergeProd$new()

PipeOpTorchMergeProd$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

mlr3torch::PipeOpTorch$shapes_out()

Method new()

Creates a new instance of this R6 class.

Usage

PipeOpTorchMergeProd$new(id = "nn_merge_prod", innum = 0, param_vals = list())

Arguments

id

(character(1))
Identifier of the resulting object.

innum

(integer(1))
The number of inputs. Default is 0 which means there is one vararg input channel.

param_vals

(list())
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction.

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpTorchMergeProd$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Construct the PipeOp
pipeop = po("nn_merge_prod")
pipeop
#> PipeOp: <nn_merge_prod> (not trained)
#> values: <list()>
#> Input channels <name [train type, predict type]>:
#>   ... [ModelDescriptor,Task]
#> Output channels <name [train type, predict type]>:
#>   output [ModelDescriptor,Task]
# The available parameters
pipeop$param_set
#> <ParamSet(0)>
#> Empty.