Base Class for Batch Normalization

Base class for batch normalization. Don't use this class directly.

Base class for batch normalization. Don't use this class directly.

Format

R6Class inheriting from PipeOpTorch, PipeOp.

R6Class inheriting from PipeOpTorchBatchNorm, PipeOpTorch, PipeOp.

R6Class inheriting from PipeOpTorchBatchNorm, PipeOpTorch, PipeOp.

R6Class inheriting from PipeOpTorchBatchNorm, PipeOpTorch, PipeOp.

Construction

PipeOpTorchBatchNorm$new(id, module_generator, min_dim, max_dim, param_vals = list())

id :: character(1)
The id for the object.

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

module_generator :: nn_module_generator
The torch module generator.

min_dim :: integer(1)
The minimum number of dimension for the input tensor.

max_dim :: integer(1)
The maximum number of dimension for the input tensor.

PipeOpTorchBatchNorm1D$new(id = "nn_batch_norm1d", param_vals = list())

id :: character(1)
The id for the object. The default is "nn_batch_norm_1d".

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

PipeOpTorchBatchNorm2D$new(id = "nn_batch_norm2d", param_vals = list())

id :: character(1)
The id for the object. The default is "nn_batch_norm_2d".

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

PipeOpTorchBatchNorm3D$new(id = "nn_batch_norm3d", param_vals = list())

id :: character(1)
The id for the object. The default is "nn_batch_norm_3d".

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

Input and Output Channels

One input channel called "input" and one output channel called"output". For an explanation see PipeOpTorch.

One input channel called "input" and one output channel called"output". For an explanation see PipeOpTorch.

One input channel called "input" and one output channel called"output". For an explanation see PipeOpTorch.

One input channel called "input" and one output channel called"output". For an explanation see PipeOpTorch.

State

The state is the value calculated by the public method shapes_out().

The state is the value calculated by the public method shapes_out().

The state is the value calculated by the public method shapes_out().

The state is the value calculated by the public method shapes_out().

Parameters

eps :: numeric(1)
A value added to the denominator for numerical stability. Default: 1e-5.

momentum :: numeric(1)
The value used for the running_mean and running_var computation. Can be set to NULL for cumulative moving average (i.e. simple average). Default: 0.1

affine :: logical(1)
a boolean value that when set to TRUE, this module has learnable affine parameters. Default: TRUE

track_running_stats :: logical(1)
a boolean value that when set to TRUE, this module tracks the running mean and variance, and when set to FALSE, this module does not track such statistics and always uses batch statistics in both training and eval modes. Default: TRUE

eps :: numeric(1)
A value added to the denominator for numerical stability. Default: 1e-5.

momentum :: numeric(1)
The value used for the running_mean and running_var computation. Can be set to NULL for cumulative moving average (i.e. simple average). Default: 0.1

affine :: logical(1)
a boolean value that when set to TRUE, this module has learnable affine parameters. Default: TRUE

track_running_stats :: logical(1)
a boolean value that when set to TRUE, this module tracks the running mean and variance, and when set to FALSE, this module does not track such statistics and always uses batch statistics in both training and eval modes. Default: TRUE

eps :: numeric(1)
A value added to the denominator for numerical stability. Default: 1e-5.

momentum :: numeric(1)
The value used for the running_mean and running_var computation. Can be set to NULL for cumulative moving average (i.e. simple average). Default: 0.1

affine :: logical(1)
a boolean value that when set to TRUE, this module has learnable affine parameters. Default: TRUE

track_running_stats :: logical(1)
a boolean value that when set to TRUE, this module tracks the running mean and variance, and when set to FALSE, this module does not track such statistics and always uses batch statistics in both training and eval modes. Default: TRUE

eps :: numeric(1)
A value added to the denominator for numerical stability. Default: 1e-5.

momentum :: numeric(1)
The value used for the running_mean and running_var computation. Can be set to NULL for cumulative moving average (i.e. simple average). Default: 0.1

affine :: logical(1)
a boolean value that when set to TRUE, this module has learnable affine parameters. Default: TRUE

track_running_stats :: logical(1)
a boolean value that when set to TRUE, this module tracks the running mean and variance, and when set to FALSE, this module does not track such statistics and always uses batch statistics in both training and eval modes. Default: TRUE

Fields

Only fields inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Only fields inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Only fields inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Only fields inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Methods

Only methods inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Only methods inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Only methods inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Only methods inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Internals

See the respective child class.

Calls torch::nn_batch_norm1d(). The parameter num_features is inferred as the second dimension of the input shape.

Calls torch::nn_batch_norm2d(). The parameter num_features is inferred as the second dimension of the input shape.

Calls torch::nn_batch_norm3d(). The parameter num_features is inferred as the second dimension of the input shape.

Credit

Parts of this documentation have been copied or adapted from the R package torch, that comes under the MIT license, which is included in the COPYRIGHTS file of the source package.

Parts of this documentation have been copied or adapted from the R package torch, that comes under the MIT license, which is included in the COPYRIGHTS file of the source package.

Parts of this documentation have been copied or adapted from the R package torch, that comes under the MIT license, which is included in the COPYRIGHTS file of the source package.

Parts of this documentation have been copied or adapted from the R package torch, that comes under the MIT license, which is included in the COPYRIGHTS file of the source package.

See also