Callback Configuration

Configures the callbacks of a deep learning model.

Format

R6Class inheriting from PipeOp.

Construction

PipeOpTorchCallbacks$new(callbacks = list(), id = "torch_callbacks", param_vals = list())

callbacks :: list of TorchCallbacks or character() or
The callbacks (or something convertible via as_torch_callbacks()). Must have unique ids. Default is list(). All callbacks are cloned during construction.

id :: character(1)
The id for the object. The default is "torch_callbacks".

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

Input and Output Channels

There is one input channel "input" and one output channel "output". During training, the channels are of class ModelDescriptor. During prediction, the channels are of class Task.

State

The state is set to an empty list().

Parameters

The parameters are defined dynamically from the callbacks, where the id of the respective callbacks is the respective set id.

Fields

Only fields inherited from PipeOp.

Methods

Only methods inherited from PipeOp.

Internals

During training the callbacks are cloned and added to the ModelDescriptor.

See also

Examples

po_cb = po("torch_callbacks", "checkpoint")
po_cb$param_set
#> <ParamSetCollection:torch_callbacks>
#>                      id    class lower upper nlevels        default value
#> 1:      checkpoint.path ParamUty    NA    NA     Inf <NoDefault[3]>      
#> 2:      checkpoint.freq ParamInt     1   Inf     Inf <NoDefault[3]>      
#> 3: checkpoint.save_last ParamLgl    NA    NA       2           TRUE      
mdin = po("torch_ingress_num")$train(list(tsk("iris")))
mdin[[1L]]$callbacks
#> named list()
mdout = po_cb$train(mdin)[[1L]]
mdout$callbacks
#> $checkpoint
#> <TorchCallback:checkpoint> Checkpoint
#> * Generator: CallbackTorchCheckpoint
#> * Parameters: list()
#> * Packages: torch,mlr3torch
#> 
# Can be called again
po_cb1 = po("torch_callbacks", t_clbk("progress"))
mdout1 = po_cb1$train(list(mdout))[[1L]]
mdout1$callbacks
#> $progress
#> <TorchCallback:progress> Progress
#> * Generator: CallbackTorchProgress
#> * Parameters: list()
#> * Packages: progress,torch,mlr3torch
#> 
#> $checkpoint
#> <TorchCallback:checkpoint> Checkpoint
#> * Generator: CallbackTorchCheckpoint
#> * Parameters: list()
#> * Packages: torch,mlr3torch
#>