Transpose 3D Convolution

Applies a 3D transposed convolution operator over an input image composed of several input planes, sometimes also called "deconvolution"

Format

R6Class inheriting from PipeOpTorchConvTranspose, PipeOpTorch, PipeOp.

Construction

PipeOpTorchConvTranspose3D$new(id = "nn_conv_transpose3d", param_vals = list())

id :: character(1)
The id for the object. The default is "nn_conv_transpose3d".

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

Input and Output Channels

One input channel called "input" and one output channel called"output". For an explanation see PipeOpTorch.

State

The state is the value calculated by the public method shapes_out().

Fields

Only fields inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Methods

Only methods inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Internals

Calls nn_conv_transpose3d. The parameter in_channels is inferred as the second dimension of the input tensor.

Credit

Parts of this documentation have been copied or adapted from the R package torch, that comes under the MIT license, which is included in the COPYRIGHTS file of the source package.

Parameters

out_channels :: integer(1)
Number of output channels produce by the convolution.

kernel_size :: integer()
Size of the convolving kernel.

stride :: integer()
Stride of the convolution. Default: 1.

padding :: integer()`
‘dilation * (kernel_size - 1) - padding’ zero-padding will be added to both sides of the input. Default: 0.

output_padding ::integer()
Additional size added to one side of the output shape. Default: 0.

groups :: integer()
Number of blocked connections from input channels to output channels. Default: 1

bias :: logical(1)
If ‘True’, adds a learnable bias to the output. Default: ‘TRUE’.

dilation :: integer()
Spacing between kernel elements. Default: 1.

padding_mode :: character(1)
The padding mode. One of "zeros", "reflect", "replicate", or "circular". Default is "zeros".

See also

Examples

obj = po("nn_conv_transpose3d", out_channels = 4, kernel_size = 2)
obj$id
#> [1] "nn_conv_transpose3d"
obj$module_generator
#> <nn_conv_transpose3d> object generator
#>   Inherits from: <inherit>
#>   Public:
#>     .classes: nn_conv_transpose3d nn_conv_transpose_nd nn_conv_nd nn_m ...
#>     initialize: function (in_channels, out_channels, kernel_size, stride = 1, 
#>     forward: function (input, output_size = NULL) 
#>     clone: function (deep = FALSE) 
#>   Parent env: <environment: 0x55f2ede45b30>
#>   Locked objects: FALSE
#>   Locked class: FALSE
#>   Portable: TRUE
obj$shapes_out(c(16, 3, 64, 64, 1000))
#> $output
#> [1]   16    4   65   65 1001
#>