Entrypoint to Torch Network

Use this as entry-point to mlr3torch-networks. Unless you are an advanced user, you should not need to use this directly but PipeOpTorchIngressNumeric, PipeOpTorchIngressCategorical or PipeOpTorchIngressImage.

Format

R6Class inheriting from PipeOp.

Construction

PipeOpTorchIngress$new(id, param_set = ps(), param_vals = list(), packages = character(0), feature_types)

id :: character(1)
The id for the object.

param_set :: link{ParamSet}{paradox}
The parameter set.

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

input :: data.table()
The input channels for this PipeOp. See PipeOp for an explanation.

output :: data.table()
The output channels for this PipeOp. See PipeOp for an explanation.

packages :: character() \cr The packages this PipeOpTorchIngress` depends on.

feature_types :: character()
The features types that can be consumed by this PipeOpTorchIngress. Must be a subset of mlr_reflections$task_feature_types.

Input and Output Channels

One input channel called "input" and one output channel called"output". For an explanation see PipeOpTorch.

State

The state is set to the input shape.

Parameters

Defined by the construction argument param_set, as well as:

select :: logical(1)
Whether to select the features of the task that can be consumed by this PipeOpTorchIngress. Default is FALSE, i.e. it errs during training if it receives a task with feature types that it cannot handle.

Fields

feature_types :: character(1)
The features types that can be consumed by this PipeOpTorchIngress.

Methods

Only methods inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Internals

Creates an object of class TorchIngressToken for the given task. The purpuse of this is to store the information on how to construct the torch dataloader from the task for this entry point of the network.

See also