Torch Entry Point for Categorical Features

Ingress PipeOp that represents a categorical (factor(), ordered() and logical()) entry point to a torch network.

Format

R6Class inheriting from PipeOpTorchIngress/PipeOpTorch.

Construction

PipeOpTorchIngressCategorical$new(id = "torch_ingress_categ", param_vals = list())

id :: character(1)
The id for the object.

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

Input and Output Channels

One input channel called "input" and one output channel called"output". For an explanation see PipeOpTorch.

State

The state is the value calculated by the public method shapes_out().

Parameters

select :: logical(1)
Whether PipeOp should selected the supported feature types. Otherwise it will err on receiving tasks with unsupported feature types.

Fields

Only fields inherited from PipeOpTorchIngress / PipeOp.

Methods

Only methods inherited from PipeOpTorchIngress / PipeOp.

Internals

Uses batchgetter_categ().

See also

Examples

po_ingress = po("torch_ingress_categ", select = TRUE)
task = tsk("german_credit")
# The output is a TorchIngressToken
token = po_ingress$train(list(task))[[1L]]
ingress = token$ingress[[1L]]
ingress$batchgetter(task$data(1, ingress$features), "cpu")
#> Error in cpp_torch_int64(): Lantern is not loaded. Please use `install_torch()` to install additional dependencies.