Torch Entry Point for Images

uses task with "imageuri" column and loads this as images. doesn't do any preprocessing or so (image resizing) and instead just errors if images don't fit. also no data augmentation etc.

Format

R6Class inheriting from PipeOpTorchIngress and PipeOp.

Construction

PipeOpTorchIngressImage$new(id = "torch_ingress_img", param_vals = list(), param_set = param_set)

Input and Output Channels

One input channel called "input" and one output channel called"output". For an explanation see PipeOpTorch.

State

The state is the value calculated by the public method shapes_out().

Parameters

select :: logical(1)
Whether PipeOp should selected the supported feature types. Otherwise it will err, when receiving tasks with unsupported feature types.

channels :: integer(1)
The number of input channels.

height :: integer(1)
The height of the pixels.

width :: integer(1)
The width of the pixels.

Fields

Only fields inherited from PipeOpTorchIngress / PipeOp.

Methods

Only methods inherited from PipeOpTorchIngress / PipeOp.

Internals

Uses magick::image_read()'to load the image.

See also

Examples

po_ingress = po("torch_ingress_img", channels = 3, height = 64, width = 64)
po_ingress
#> PipeOp: <torch_ingress_img> (not trained)
#> values: <channels=3, height=64, width=64, select=FALSE>
#> Input channels <name [train type, predict type]>:
#>   input [Task,Task]
#> Output channels <name [train type, predict type]>:
#>   output [ModelDescriptor,Task]