Loss Configuration

Configures the loss of a deep learning model.

Format

R6Class inheriting from PipeOp.

Construction

PipeOpTorchLoss$new(loss, id = "torch_loss", param_vals = list())

loss :: TorchLoss or character(1) or nn_loss
The loss (or something convertible via as_torch_loss()). This object is cloned during construction.

id :: character(1)
The id for the object. The default is "torch_loss".

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

Input and Output Channels

There is one input channel "input" and one output channel "output". During training, the channels are of class ModelDescriptor. During prediction, the channels are of class Task.

State

The state is set to an empty list().

Parameters

The parameters are defined dynamically from the loss set during construction.

Fields

Only fields inherited from PipeOp.

Methods

Only methods inherited from PipeOp.

Internals

During training the loss is cloned and added to the ModelDescriptor.

See also

Examples

po_loss = po("torch_loss", "cross_entropy")
po_loss$param_set
#> <ParamSet:torch_loss>
#>              id    class lower upper nlevels default value
#> 1:       weight ParamUty    NA    NA     Inf              
#> 2: ignore_index ParamInt  -Inf   Inf     Inf    -100      
#> 3:    reduction ParamFct    NA    NA       2    mean      
mdin = po("torch_ingress_num")$train(list(tsk("iris")))
mdin[[1L]]$loss
#> NULL
mdout = po_loss$train(mdin)[[1L]]
mdout$loss
#> <TorchLoss:cross_entropy> Cross Entropy
#> * Generator: nn_crossentropy_loss
#> * Parameters: list()
#> * Packages: torch,mlr3torch
#> * Task Types: classif