Merge by Product

Calculates the product of all input tensors.

Format

R6Class object inheriting from PipeOp/PipeOpTorch/PipeOpTorchMerge.

Construction

PipeOpTorchMergeProd$new(id = "nn_merge_prod", innum = 0, param_vals = list())

id :: character(1)
The id for the object. The default is "nn_merge_prod".

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

innum :: integer(1)
The number of inputs. Default is 0 which means there is one vararg input channel.

State

The state is the value calculated by the public method shapes_out().

Parameters

No parameters.

Fields

Only fields inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Methods

Only methods inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Internals

Calls nn_merge_prod() when trained.

Input and Output Channels

PipeOpTorchMerges has either a vararg input channel if the constructor argument innum is not set, or input channels "input1", ..., "input<innum>". There is one output channel "output". For an explanation see PipeOpTorch.

See also

Examples

obj = po("nn_merge_prod")
obj$id
#> [1] "nn_merge_prod"
obj$module_generator
#> <nn_merge_prod> object generator
#>   Inherits from: <inherit>
#>   Public:
#>     .classes: nn_merge_prod nn_module
#>     initialize: function () 
#>     forward: function (...) 
#>     clone: function (deep = FALSE) 
#>   Parent env: <environment: 0x55f2ed5ea8e8>
#>   Locked objects: FALSE
#>   Locked class: FALSE
#>   Portable: TRUE
obj$shapes_out(list(input1 = c(16, 5, 5), input2 = c(16, 5, 5)))
#> $output
#> [1] 16  5  5
#>