RReLU Activation Function

Randomized leaky ReLU.

Format

R6Class inheriting from PipeOpTorch, PipeOp.

Construction

PipeOpTorchRReLU$new(id = "nn_rrelu", param_vals = list())

id :: character(1)
The id for the object. The default is "nn_rrelu".

param_vals :: named list()
List of hyperparameter settings to overwrite the initial values. Default is list().

Input and Output Channels

One input channel called "input" and one output channel called"output". For an explanation see PipeOpTorch.

State

The state is the value calculated by the public method shapes_out().

Parameters

lower:: numeric(1)
Lower bound of the uniform distribution. Default: 1/8.

upper:: numeric(1)
Upper bound of the uniform distribution. Default: 1/3.

inplace :: logical(1)
Whether to do the operation in-place. Default: FALSE.

Fields

Only fields inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Methods

Only methods inherited from PipeOpTorch/PipeOp{mlr3pipelines}].

Internals

Calls torch::nn_rrelu() when trained.

Credit

Parts of this documentation have been copied or adapted from the R package torch, that comes under the MIT license, which is included in the COPYRIGHTS file of the source package.

See also