Adjust Gamma Transformation

Calls torchvision::transform_adjust_gamma, see there for more information on the parameters. The preprocessing is applied to each element of a batch individually.

Format

R6Class inheriting from PipeOpTaskPreprocTorch.

Construction

po("trafo_adjust_gamma"")

Parameters

Id	Type	Default	Levels	Range

gamma	numeric	-		\([0, \infty)\)

gain	numeric	1		\((-\infty, \infty)\)

stages	character	-	train, predict, both	-

affect_columns	untyped	selector_all()		-