Resampling Using Row Roles

In mlr3 it is possible to manually set the row roles use and test. This resampling sets the rows from use as the training ids and those from test as the test ids. This can be useful for tasks like tiny imagenet that come with predefined splits.

Super class

mlr3::Resampling -> ResamplingRowRoles

Active bindings

iters

(integer(1))
The number of iterations which are always one for this resampling.

Methods

Public methods

ResamplingRowRoles$new()

ResamplingRowRoles$instantiate()

ResamplingRowRoles$train_set()

ResamplingRowRoles$test_set()

ResamplingRowRoles$clone()

Inherited methods

mlr3::Resampling$format()

mlr3::Resampling$help()

mlr3::Resampling$print()

Method new()

Creates a new instance of this R6 class.

Usage

ResamplingRowRoles$new()

Method instantiate()

Materializes fixed training and test splits for a given task and stores them in $instance.

Usage

ResamplingRowRoles$instantiate(task)

Arguments

task

(Task)
Task used for instantiation.

Returns

Returns the object itself, but modified by reference. You need to explicitly $clone() the object beforehand if you want to keeps the object in its previous state.

Method train_set()

Returns the row ids of the i-th train set.

Usage

ResamplingRowRoles$train_set(i)

Arguments

i

(integer(1))
Iteration.

Returns

(integer()) of row ids.

Method test_set()

Returns the row ids of the i-th test set.

Usage

ResamplingRowRoles$test_set(i)

Arguments

i

(integer(1))
Iteration.

Returns

(integer()) of row ids.

Method clone()

The objects of this class are cloneable with this method.

Usage

ResamplingRowRoles$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

resampling = rsmp("row_roles")
resampling
#> <ResamplingRowRoles>: Row Roles
#> * Iterations: 1
#> * Instantiated: FALSE
#> * Parameters: list()

task = tsk("mtcars")
splits = partition(task)
task$row_roles$use = splits$train
task$row_roles$test = splits$test
rr = resample(task, lrn("regr.featureless"), resampling)
rr$score()
#>    task_id       learner_id resampling_id iteration regr.mse
#>     <char>           <char>        <char>     <int>    <num>
#> 1:  mtcars regr.featureless     row_roles         1 32.37742
#> Hidden columns: task, learner, resampling, prediction