Tiny ImageNet Classification Task

Subset of the famous ImageNet dataset. The data is obtained from torchvision::tiny_imagenet_dataset(). It contains the train, validation and test data. The row role use is set to the training data, the row role test to the valdiation data and the row role holdout is set to the test data. There are no labels available for the test data.

The underlying DataBackend contains columns "class", "image", "row_id", "split", where the last column indicates whether the row belongs to the train, validation or test set.

Construction

tsk("tiny_imagenet")

Meta Information

Task type: “classif”

Properties: “multiclass”

Has Missings: no

Target: “class”

Features: “image”

Backend Dimension: 120000x3

Default Roles (use / test / holdout): 100000, 10000, 10000

References

Deng, Jia, Dong, Wei, Socher, Richard, Li, Li-Jia, Li, Kai, Fei-Fei, Li (2009). “Imagenet: A large-scale hierarchical image database.” In 2009 IEEE conference on computer vision and pattern recognition, 248--255. IEEE.

Examples

task = tsk("tiny_imagenet")
task
#> <TaskClassif:tiny_imagenet> (100000 x 2): ImageNet Subset
#> * Target: class
#> * Properties: multiclass
#> * Features (1):
#>   - lt (1): image
#> * Additional Row Roles: 10000 (test), 10000 (holdout)