GeGLU Module

This module implements the Gaussian Error Linear Unit Gated Linear Unit (GeGLU) activation function. It computes \(\text{GeGLU}(x, g) = x \cdot \text{GELU}(g)\) where \(x\) and \(g\) are created by splitting the input tensor in half along the last dimension.

Usage

nn_geglu()

References

Shazeer N (2020). “GLU Variants Improve Transformer.” 2002.05202, https://arxiv.org/abs/2002.05202.

Examples

x = torch::torch_randn(10, 10)
glu = nn_geglu()
glu(x)
#> torch_tensor
#> -1.4820e-01 -9.0112e-01  2.0016e-02  1.0636e+00  5.8085e-01
#>  1.9725e-02  5.0573e-01  2.5896e-01  4.2649e-02  7.0197e-01
#>  8.1926e-01 -3.8973e+00 -1.3084e-02 -2.7707e+00  3.6328e-01
#>  8.0580e-05  7.3979e-02  5.0772e-03 -1.1607e+00  7.5161e-01
#> -2.1581e-01 -1.3747e-03  9.1735e-02  1.8623e-01  2.7342e-01
#>  1.0281e-01  2.2399e-01  2.6788e-03 -1.1555e-01 -7.0723e-01
#> -1.8337e-01 -6.9767e-02 -5.4613e-01 -1.4451e-01 -5.8867e-01
#> -1.5937e-02  5.4242e-02  2.8106e-01  1.6854e-01  9.3570e-01
#> -3.2754e+00 -1.5075e-01 -2.4521e-02  7.4320e-02  2.3309e-01
#> -3.8158e-01 -1.0168e+00  4.8860e-03  1.0024e-01  2.8787e-01
#> [ CPUFloatType{10,5} ]