Create a Torch Callback

Convenience function to create a custom callback for torch. For more information on how to correctly implement a new callback, see CallbackTorch. It returns a TorchCallback that wrapping a CallbackTorch.

torch_callback(
  id,
  classname = paste0("CallbackTorch", capitalize(id)),
  param_set = NULL,
  packages = NULL,
  label = capitalize(id),
  man = NULL,
  on_begin = NULL,
  on_end = NULL,
  on_epoch_begin = NULL,
  on_before_valid = NULL,
  on_epoch_end = NULL,
  on_batch_begin = NULL,
  on_batch_end = NULL,
  on_after_backward = NULL,
  on_batch_valid_begin = NULL,
  on_batch_valid_end = NULL,
  public = NULL,
  private = NULL,
  active = NULL,
  parent_env = parent.frame(),
  inherit = CallbackTorch
)

Arguments

id

(character(1))
`
The id for the callbacks. Note that the ids of callbacks passed to a learner must be unique.

classname

(character(1))
The class name.

param_set

(ParamSet)
The parameter set, if not present it is inferred from the initialize method passed through the public function.

packages

(character())
The packages the callback depends on. Default isNULL`.

label

(character(1))
Label for the new instance.

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

on_begin, on_end, on_epoch_begin, on_before_valid, on_epoch_end, on_batch_begin, on_batch_end, on_after_backward, on_batch_valid_begin, on_batch_valid_end

(function)
Function to execute at the given stage, see section Stages.

public, private, active

(list())
Additional public, private, and active fields to add to the callback.

parent_env

(environment())
The parent environment for the R6Class.

inherit

(R6ClassGenerator)
From which class to inherit. This class must either be CallbackTorch (default) or inherit from it.

Value

TorchCallback

Internals

It first creates an R6ClassGenerator using torch_callbackand when wraps this generator in a TorchCallback.

Stages

begin :: Run before the training loop begins.

epoch_begin :: Run he beginning of each epoch.

before_validation :: Run before each validation loop.

batch_begin :: Run before the forward call.

after_backward :: Run after the backward call.

batch_end :: Run after the optimizer step.

batch_valid_begin :: Run before the forward call in the validation loop.

batch_valid_end :: Run after the forward call in the validation loop.

epoch_end :: Run at the end of each epoch.

end :: Run at last, using on.exit().

See also

Examples

custom_tcb = torch_callback(
  "mycallback",
  public = list(
    initialize = function(greeting) {
      self$greeting = greeting
    }
  ),
  on_begin = function(ctx) {
    cat(self$greeting, ctx$name, "\n")
  },
  on_end = function(ctx) {
    cat("Bye", ctx$name, "\n")
  }
)

custom_tcb$param_set$set_values(greeting = "Wazzuuup")

cb = custom_tcb$generate()

ctx = new.env()
ctx$name = "Albert"

f = function(ctx, cb) {
  cb$on_begin(ctx)
  catn("Doing heavy work ...")
  cb$on_end(ctx)
}

f(ctx, cb)
#> Wazzuuup Albert 
#> Doing heavy work ...
#> Bye Albert