Base Class for Torch Wrappers

Abstract Base Class from which TorchLoss, TorchOptimizer, and TorchCallback inherit. This class wraps a generator (R6Class Generator or the torch version of such a generator) and annotates it.

Format

R6Class

Arguments

man

(character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Construction

TorchWrapper$new(generator, id, param_set = NULL, packages = NULL, label = id, man = NULL)

generator :: function or R6ClassGenerator
The wrapped generator that is described.

id :: character(1)
The identifier of the object. Used to e.g. retrieve it from a dictionary.

param_set :: paradox::ParamSet
The parameter set that describes the arguments of the generator.

packages :: character()
The packages the generator depends on.

label :: character(1)
The label, which is used for printing. Defaults to id.

man :: (character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object.

Parameters

Defined by the constructor argument param_set.

Fields

generator :: function or R6ClassGenerator
The wrapped generator that is described.

id :: character(1)
The identifier of the object. Used to e.g. retrieve it from a dictionary.

param_set :: paradox::ParamSet
The parameter set that describes the arguments of the generator.

packages :: character()
The packages the generator depends on.

label :: character(1)
The label, which is used for printing. Defaults to id.

man :: (character(1))
String in the format [pkg]::[topic] pointing to a manual page for this object. The referenced help package can be opened via method $help().

Methods

generate()
() -> any Calls the generator with the given parameter values.

print(...)
() -> CallbackTorch Prints the object.

help()
() -> help file
Displays the help file of the wrapped object.

Internals

When a deep clone is created, the wrapped generator is not cloned.

See also