Materialize Lazy Tensor Columns

This will materialize a lazy_tensor() or a data.frame() / list() containing -- among other things -- lazy_tensor() columns. I.e. the data described in the underlying DataDescriptors is loaded for the indices in the lazy_tensor(), is preprocessed and then put unto the specified device. Because not all elements in a lazy tensor must have the same shape, a list of tensors is returned by default. If all elements have the same shape, these tensors can also be rbinded into a single tensor (parameter rbind).

Usage

materialize(x, device = "cpu", rbind = FALSE, ...)

# S3 method for list
materialize(x, device = "cpu", rbind = FALSE, cache = "auto", ...)

Arguments

x

(any)
The object to materialize. Either a lazy_tensor or a list() / data.frame() containing lazy_tensor columns.

device

(character(1))
The torch device.

rbind

(logical(1))
Whether to rbind the lazy tensor columns (TRUE) or return them as a list of tensors (FALSE). In the second case, there is no batch dimension.

...

(any)
Additional arguments.

cache

(character(1) or environment() or NULL)
Optional cache for (intermediate) materialization results. Per default, caching will be enabled when the same dataset or data descriptor (with different output pointer) is used for more than one lazy tensor column.

Value

(list() of lazy_tensors or a lazy_tensor)

Details

Materializing a lazy tensor consists of:

Loading the data from the internal dataset of the DataDescriptor.

Processing these batches in the preprocessing Graphs.

Returning the result of the PipeOp pointed to by the DataDescriptor (pointer).

With multiple lazy_tensor columns we can benefit from caching because: a) Output(s) from the dataset might be input to multiple graphs. b) Different lazy tensors might be outputs from the same graph.

For this reason it is possible to provide a cache environment. The hash key for a) is the hash of the indices and the dataset. The hash key for b) is the hash of the indices, dataset and preprocessing graph.

Examples

lt1 = as_lazy_tensor(torch_randn(10, 3))
materialize(lt1, rbind = TRUE)
#> torch_tensor
#>  1.0590 -0.3944 -1.0170
#> -0.1248  0.3669  0.0836
#>  1.2008  0.4891 -1.0015
#> -0.3376  1.3441  0.2368
#>  0.2465  0.1548  1.1804
#> -0.1316  0.8191 -1.1065
#> -0.2605 -0.2690  1.7256
#>  0.7394  0.0286  0.5360
#>  1.8253 -1.9279  0.7284
#>  0.2973 -0.4586  0.9945
#> [ CPUFloatType{10,3} ]
materialize(lt1, rbind = FALSE)
#> [[1]]
#> torch_tensor
#>  1.0590
#> -0.3944
#> -1.0170
#> [ CPUFloatType{3} ]
#> 
#> [[2]]
#> torch_tensor
#> -0.1248
#>  0.3669
#>  0.0836
#> [ CPUFloatType{3} ]
#> 
#> [[3]]
#> torch_tensor
#>  1.2008
#>  0.4891
#> -1.0015
#> [ CPUFloatType{3} ]
#> 
#> [[4]]
#> torch_tensor
#> -0.3376
#>  1.3441
#>  0.2368
#> [ CPUFloatType{3} ]
#> 
#> [[5]]
#> torch_tensor
#>  0.2465
#>  0.1548
#>  1.1804
#> [ CPUFloatType{3} ]
#> 
#> [[6]]
#> torch_tensor
#> -0.1316
#>  0.8191
#> -1.1065
#> [ CPUFloatType{3} ]
#> 
#> [[7]]
#> torch_tensor
#> -0.2605
#> -0.2690
#>  1.7256
#> [ CPUFloatType{3} ]
#> 
#> [[8]]
#> torch_tensor
#>  0.7394
#>  0.0286
#>  0.5360
#> [ CPUFloatType{3} ]
#> 
#> [[9]]
#> torch_tensor
#>  1.8253
#> -1.9279
#>  0.7284
#> [ CPUFloatType{3} ]
#> 
#> [[10]]
#> torch_tensor
#>  0.2973
#> -0.4586
#>  0.9945
#> [ CPUFloatType{3} ]
#> 
lt2 = as_lazy_tensor(torch_randn(10, 4))
d = data.table::data.table(lt1 = lt1, lt2 = lt2)
materialize(d, rbind = TRUE)
#> $lt1
#> torch_tensor
#>  1.0590 -0.3944 -1.0170
#> -0.1248  0.3669  0.0836
#>  1.2008  0.4891 -1.0015
#> -0.3376  1.3441  0.2368
#>  0.2465  0.1548  1.1804
#> -0.1316  0.8191 -1.1065
#> -0.2605 -0.2690  1.7256
#>  0.7394  0.0286  0.5360
#>  1.8253 -1.9279  0.7284
#>  0.2973 -0.4586  0.9945
#> [ CPUFloatType{10,3} ]
#> 
#> $lt2
#> torch_tensor
#>  0.0279 -1.1844  0.7732  0.1019
#>  0.1923  0.2514  0.8052 -0.3557
#>  0.1369 -0.0092 -1.9954  0.8914
#> -0.0198  0.7666 -0.0419 -1.5322
#> -0.6783 -0.4461  1.3078  1.6142
#>  0.4490 -1.7187  0.6425  0.4901
#> -0.9419 -0.2317 -0.2003  0.5104
#>  2.1624 -0.9531 -1.0710  0.6599
#>  0.9902 -0.3358 -0.9807 -2.7776
#>  0.2460  0.8132  1.4297  0.7715
#> [ CPUFloatType{10,4} ]
#> 
materialize(d, rbind = FALSE)
#> $lt1
#> $lt1[[1]]
#> torch_tensor
#>  1.0590
#> -0.3944
#> -1.0170
#> [ CPUFloatType{3} ]
#> 
#> $lt1[[2]]
#> torch_tensor
#> -0.1248
#>  0.3669
#>  0.0836
#> [ CPUFloatType{3} ]
#> 
#> $lt1[[3]]
#> torch_tensor
#>  1.2008
#>  0.4891
#> -1.0015
#> [ CPUFloatType{3} ]
#> 
#> $lt1[[4]]
#> torch_tensor
#> -0.3376
#>  1.3441
#>  0.2368
#> [ CPUFloatType{3} ]
#> 
#> $lt1[[5]]
#> torch_tensor
#>  0.2465
#>  0.1548
#>  1.1804
#> [ CPUFloatType{3} ]
#> 
#> $lt1[[6]]
#> torch_tensor
#> -0.1316
#>  0.8191
#> -1.1065
#> [ CPUFloatType{3} ]
#> 
#> $lt1[[7]]
#> torch_tensor
#> -0.2605
#> -0.2690
#>  1.7256
#> [ CPUFloatType{3} ]
#> 
#> $lt1[[8]]
#> torch_tensor
#>  0.7394
#>  0.0286
#>  0.5360
#> [ CPUFloatType{3} ]
#> 
#> $lt1[[9]]
#> torch_tensor
#>  1.8253
#> -1.9279
#>  0.7284
#> [ CPUFloatType{3} ]
#> 
#> $lt1[[10]]
#> torch_tensor
#>  0.2973
#> -0.4586
#>  0.9945
#> [ CPUFloatType{3} ]
#> 
#> 
#> $lt2
#> $lt2[[1]]
#> torch_tensor
#>  0.0279
#> -1.1844
#>  0.7732
#>  0.1019
#> [ CPUFloatType{4} ]
#> 
#> $lt2[[2]]
#> torch_tensor
#>  0.1923
#>  0.2514
#>  0.8052
#> -0.3557
#> [ CPUFloatType{4} ]
#> 
#> $lt2[[3]]
#> torch_tensor
#>  0.1369
#> -0.0092
#> -1.9954
#>  0.8914
#> [ CPUFloatType{4} ]
#> 
#> $lt2[[4]]
#> torch_tensor
#> -0.0198
#>  0.7666
#> -0.0419
#> -1.5322
#> [ CPUFloatType{4} ]
#> 
#> $lt2[[5]]
#> torch_tensor
#> -0.6783
#> -0.4461
#>  1.3078
#>  1.6142
#> [ CPUFloatType{4} ]
#> 
#> $lt2[[6]]
#> torch_tensor
#>  0.4490
#> -1.7187
#>  0.6425
#>  0.4901
#> [ CPUFloatType{4} ]
#> 
#> $lt2[[7]]
#> torch_tensor
#> -0.9419
#> -0.2317
#> -0.2003
#>  0.5104
#> [ CPUFloatType{4} ]
#> 
#> $lt2[[8]]
#> torch_tensor
#>  2.1624
#> -0.9531
#> -1.0710
#>  0.6599
#> [ CPUFloatType{4} ]
#> 
#> $lt2[[9]]
#> torch_tensor
#>  0.9902
#> -0.3358
#> -0.9807
#> -2.7776
#> [ CPUFloatType{4} ]
#> 
#> $lt2[[10]]
#> torch_tensor
#>  0.2460
#>  0.8132
#>  1.4297
#>  0.7715
#> [ CPUFloatType{4} ]
#> 
#>