Torch Entry Point for Images

uses task with "imageuri" column and loads this as images. doesn't do any preprocessing or so (image resizing) and instead just errors if images don't fit. also no data augmentation etc.

Parameters

select :: logical(1)
Whether PipeOp should selected the supported feature types. Otherwise it will err, when receiving tasks with unsupported feature types.

channels :: integer(1)
The number of input channels.

height :: integer(1)
The height of the pixels.

width :: integer(1)
The width of the pixels.

Internals

Uses magick::image_read() to load the image.

Input and Output Channels

One input channel called "input" and one output channel called "output". For an explanation see PipeOpTorch.

State

The state is set to the input shape.

See also

Super classes

mlr3pipelines::PipeOp -> mlr3torch::PipeOpTorchIngress -> PipeOpTorchIngressImage

Methods

Public methods

PipeOpTorchIngressImage$new()

PipeOpTorchIngressImage$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Creates a new instance of this R6 class.

Usage

PipeOpTorchIngressImage$new(id = "torch_ingress_img", param_vals = list())

Arguments

id

(character(1))
Identifier of the resulting object.

param_vals

(list())
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction.

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpTorchIngressImage$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

po_ingress = po("torch_ingress_img", channels = 3, height = 64, width = 64)
po_ingress
#> PipeOp: <torch_ingress_img> (not trained)
#> values: <channels=3, height=64, width=64>
#> Input channels <name [train type, predict type]>:
#>   input [Task,Task]
#> Output channels <name [train type, predict type]>:
#>   output [ModelDescriptor,Task]