Ingress for Lazy Tensor

Ingress for a single lazy_tensor column.

Parameters

shape :: integer()
The shape of the tensor, where the first dimension (batch) must be NA. When it is not specified, the lazy tensor input column needs to have a known shape.

Internals

The returned batchgetter materializes the lazy tensor column to a tensor.

Input and Output Channels

One input channel called "input" and one output channel called "output". For an explanation see PipeOpTorch.

State

The state is set to the input shape.

See also

Super classes

mlr3pipelines::PipeOp -> mlr3torch::PipeOpTorchIngress -> PipeOpTorchIngressLazyTensor

Methods

Public methods

PipeOpTorchIngressLazyTensor$new()

PipeOpTorchIngressLazyTensor$clone()

Inherited methods

mlr3pipelines::PipeOp$help()

mlr3pipelines::PipeOp$predict()

mlr3pipelines::PipeOp$print()

mlr3pipelines::PipeOp$train()

Method new()

Creates a new instance of this R6 class.

Usage

PipeOpTorchIngressLazyTensor$new(
  id = "torch_ingress_ltnsr",
  param_vals = list()
)

Arguments

id

(character(1))
Identifier of the resulting object.

param_vals

(list())
List of hyperparameter settings, overwriting the hyperparameter settings that would otherwise be set during construction.

Method clone()

The objects of this class are cloneable with this method.

Usage

PipeOpTorchIngressLazyTensor$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

po_ingress = po("torch_ingress_ltnsr")
task = tsk("lazy_iris")

md = po_ingress$train(list(task))[[1L]]

ingress = md$ingress
x_batch = ingress[[1L]]$batchgetter(data = task$data(1, "x"), device = "cpu", cache = NULL)
x_batch
#> torch_tensor
#>  5.1000  3.5000  1.4000  0.2000
#> [ CPUFloatType{1,4} ]

# Now we try a lazy tensor with unknown shape, i.e. the shapes between the rows can differ

ds = dataset(
  initialize = function() self$x = list(torch_randn(3, 10, 10), torch_randn(3, 8, 8)),
  .getitem = function(i) list(x = self$x[[i]]),
  .length = function() 2)()

task_unknown = as_task_regr(data.table(
  x = as_lazy_tensor(ds, dataset_shapes = list(x = NULL)),
  y = rnorm(2)
), target = "y", id = "example2")

# this task (as it is) can NOT be processed by PipeOpTorchIngressLazyTensor
# It therefore needs to be preprocessed
po_resize = po("trafo_resize", size = c(6, 6))
task_unknown_resize = po_resize$train(list(task_unknown))[[1L]]

# printing the transformed column still shows unknown shapes,
# because the preprocessing pipeop cannot infer them,
# however we know that the shape is now (3, 10, 10) for all rows
task_unknown_resize$data(1:2, "x")
#>                x
#>    <lazy_tensor>
#> 1:      <tnsr[]>
#> 2:      <tnsr[]>
po_ingress$param_set$set_values(shape = c(NA, 3, 6, 6))

md2 = po_ingress$train(list(task_unknown_resize))[[1L]]

ingress2 = md2$ingress
x_batch2 = ingress2[[1L]]$batchgetter(
  data = task_unknown_resize$data(1:2, "x"),
  device = "cpu",
  cache = NULL
)

x_batch2
#> torch_tensor
#> (1,1,.,.) = 
#>  -0.5772  0.3215 -0.1803 -1.2717 -0.5904 -1.4720
#>  -1.0981  0.8400 -0.8919 -0.6404  1.0732 -1.1724
#>  -0.2378  1.2242 -0.5522  0.9572 -0.2218  0.3394
#>  -0.6597  0.8500  0.1286 -0.4021  1.4033  0.8949
#>   0.0724  0.2071 -0.3106  0.4994 -2.4397 -0.7842
#>  -0.6800 -0.3626  0.0915  0.4442 -1.3872  1.0929
#> 
#> (2,1,.,.) = 
#>   1.4872 -0.2433 -0.7385 -0.4606  0.2983  0.3474
#>   0.0553  0.3229 -0.4424 -1.1494 -1.2382 -0.9232
#>  -0.3575 -0.5874 -0.9566  0.5998  0.5222  0.6538
#>   0.0882 -0.3502  0.8155  0.2002  0.8079  1.6346
#>  -0.2565 -0.5153  0.1198  0.5984  0.1401  0.2278
#>  -0.9837  0.5641  0.5496  0.1505 -0.5780  1.7939
#> 
#> (1,2,.,.) = 
#>  -0.0656 -0.3430 -0.4046  0.2551  0.7116  0.1056
#>  -1.2322 -0.2117 -0.2795 -0.5665 -1.0839  0.3401
#>  -0.1552  0.9564  0.2326 -0.6902  1.5751  0.2677
#>   0.0601  0.3894  0.3607  0.0486 -0.3460 -0.5240
#>   0.2955 -1.2738  0.8586  0.1892 -1.1308 -0.9294
#>   1.0110  0.6762 -0.5201  0.9524 -0.6527 -0.8129
#> 
#> (2,2,.,.) = 
#>  -0.0983  0.2636  0.1246 -0.5797 -0.6730 -1.3052
#>  -0.5175  0.8516  0.3581 -0.9017  0.4513  0.4203
#>  -0.4993  0.3641  0.6648 -0.3858  0.8640  0.0686
#>   0.2410 -0.4793  0.5258  0.4028 -0.1881 -0.3267
#>   0.8562  0.0953  0.6797 -0.2245 -0.3493 -0.2846
#> ... [the output was truncated (use n=-1 to disable)]
#> [ CPUFloatType{2,3,6,6} ]