GeGLU Module

This module implements the Gaussian Error Linear Unit Gated Linear Unit (GeGLU) activation function. It computes \(\text{GeGLU}(x, g) = x \cdot \text{GELU}(g)\) where \(x\) and \(g\) are created by splitting the input tensor in half along the last dimension.

Usage

nn_geglu()

References

Shazeer N (2020). “GLU Variants Improve Transformer.” 2002.05202, https://arxiv.org/abs/2002.05202.

Examples

x = torch::torch_randn(10, 10)
glu = nn_geglu()
glu(x)
#> torch_tensor
#> -0.0584  0.6158 -0.4384  0.0280  0.2626
#> -0.7047  0.1258 -0.2079 -0.0699  0.0103
#>  0.0580 -0.1302 -0.4988 -1.5453 -0.0945
#> -0.0444 -0.1024 -0.0405 -0.3091  0.0541
#> -0.0903  0.0610 -0.0151 -0.0148 -0.0556
#> -0.1326 -1.3856 -1.6285  0.3681  0.1110
#>  3.5029  0.2076 -0.6980  0.2473 -0.2352
#>  2.3749 -0.0268  0.2290  0.1406 -0.0542
#> -0.0477 -0.0065 -0.0776 -0.0240 -0.0235
#> -0.9742 -0.4138 -0.0002  0.0126  0.0193
#> [ CPUFloatType{10,5} ]